VolatileDB: move modules under Impl
