+++ ouroboros-consensus/src/Ouroboros/Consensus/Storage/VolatileDB/Impl/State.hs
@@ -33,10 +33,10 @@ import           Ouroboros.Consensus.Util.IOLike
 import           Ouroboros.Consensus.Storage.FS.API
 import           Ouroboros.Consensus.Storage.FS.API.Types
 import           Ouroboros.Consensus.Storage.VolatileDB.API
-import qualified Ouroboros.Consensus.Storage.VolatileDB.FileInfo as FileInfo
-import           Ouroboros.Consensus.Storage.VolatileDB.Index (Index)
-import qualified Ouroboros.Consensus.Storage.VolatileDB.Index as Index
-import           Ouroboros.Consensus.Storage.VolatileDB.Util
+import qualified Ouroboros.Consensus.Storage.VolatileDB.Impl.FileInfo as FileInfo
+import           Ouroboros.Consensus.Storage.VolatileDB.Impl.Index (Index)
+import qualified Ouroboros.Consensus.Storage.VolatileDB.Impl.Index as Index
+import           Ouroboros.Consensus.Storage.VolatileDB.Impl.Util
 
 {------------------------------------------------------------------------------
   Main types
